Move data management test from integration to a unit test.
